Updates
o Refactored code with ChatGPT.
o Updated posts to their second draft.
o Created new post for troubleshooting and project management.
o Reworked the loader so I don't have to keep updating it, it now will try to load the pages if they exist by the number of posts specified in the launcher.